backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
bcrypt==4.0.1
voluptuous==0.13.1
Flask-Mail==0.9.1
PyNaCl==1.5.0
psycopg2
python-dateutil==2.8.2
python-slugify==8.0.1
requests==2.28.2
PyYAML==6.0
